% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishLength.R
\name{calc_avg_length}
\alias{calc_avg_length}
\title{Calculate the average length of the fish}
\usage{
calc_avg_length(lengths)
}
\arguments{
\item{lengths}{A numeric vector of fish lengths}
}
\value{
The average length of the fish
}
\description{
This function takes a numeric vector of fish lengths and returns their average.
}
\examples{
calc_avg_length(c(5, 10, 15))
}
