% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishWeight.R
\name{calc_avg_weight}
\alias{calc_avg_weight}
\title{Calculate the average weight of the fish}
\usage{
calc_avg_weight(weights)
}
\arguments{
\item{weights}{A numeric vector of fish weights}
}
\value{
The average weight of the fish
}
\description{
This function takes a numeric vector of fish weights and returns their average.
}
\examples{
calc_avg_weight(c(0.5, 1.2, 2.4))
}
