% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishValue.R
\name{calc_tank_value}
\alias{calc_tank_value}
\title{Calculate the wholesale value of the tank}
\usage{
calc_tank_value(avg_length, tank_value, weights)
}
\arguments{
\item{avg_length}{A numeric value for the average length of the fish}

\item{tank_value}{A numeric value for the wholesale value of the tank}

\item{weights}{A numeric vector of fish weights}
}
\value{
The wholesale value of the tank
}
\description{
This function takes a numeric value for the average length of the fish and the wholesale value of the tank, and returns the wholesale value of the tank based on the length and weight of the fish.
}
\examples{
calc_tank_value(10, 100, c(0.5, 1.2, 2.4))
}
